[op][upsable][a11y][aria] Force focus to accessible item on mouse move.
:Release Notes:
Update focus on mouse hover so it is read out, required for touch accessibility exploration.

:Detailed Notes:
This patch to support feature called 'Explore by touch in TalkBack'
for accessibility. As you move your finger over the screen, TalkBack
announces what's under your finger.

For this feature, we force to focus on the item under the finger in
web engine. Touch events are converted to mouse move events in other
component when accessibility feature is on. So in web engine, it
only needs to handle mouse event instead of touch event when
accessibility feature is on.

The feature is enabled with command line option
"--accessibility-explore-by-mouse" and when accessibility read out
feature is on at the settings menu.

In the implementation, when m_nodeUnderMouse is null or is different
with current target node, then call handleMouseFocus()
to force focus on current target node.

:Testing Performed:
1. Build tested
2. Enable audio guidance using luna command below,
luna-send -f -n 1 luna://com.lge.settingsservice/setSystemSettings
'{"category":"option", "settings":{"audioGuidance":"on"}}'
3. Verifed accessible i.e focussible events trigger as we move mouse
pointer on the screen using logs added,
@AXObjectCacheImpl::postPlatformNotification()

:QA Notes:
Test-cases: GT-26435

:Issues Addressed:
[WRN-4341] Port : webOS a11y

Change-Id: I733c6a3ad4fc7bd0aef4547c963812a03674f676
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/316128
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/317648
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>