[op][upsable][a11y] Do not forget AXEventGenerator reparented pending events
:Release Notes:
Preserve pending accessible events when a node is reparented.

:Detailed Notes:
In case a node had pending event generator events (as live region
events or alerts), and that node would be reparented before the
event was processed by BrowserAccessibilityManager, then set of
pending events would be cleared before they would be processed.

This change stores temporally the list of pending events for a node
in a new map indexed by the node ID. Node ID is fine for reparenting
as even if the node is recreated, the ID will be the same. Once the
node is reparented then we readd the events to the node (that could
have changed).

:Testing Performed:
1. Enable audio guidance.
2. Launch qa-a11y
3. Launch VideoPlayer example -> TTS should say "Loading" and then
the title of the video. Without this patch, it would simply say
the title.

:QA Notes:
Test-cases: GT-23902, GT-24808, GT-25165, GT-25213, GT-25353, GT-25512

:Issues Addressed:
[PLAT-84575] [Enact A11y] device is not reading out as expected in video player menu.
[WRN-4341] Port : webOS a11y

Change-Id: Iceb87f6c38573d00acde347cc45197978369cd24
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/316130
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/317650
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>