[op][n_upsable][dynamic_sized_window] Implement dynamic sized window
:Release Notes:
Implement dynamic sized window.
Upscale video to fullscreen providing additional
contents scale to media stack. Keep size of
size-overriding apps when relaunched.

:Detailed Notes:
If the application's actual resolution is different
from platform resolution, then WAM & webOS Webview
adjusts app's content size to fit the wayland window.
It, however, is redundant to create a larger window
than app's resolution, in terms of WAM's memory usage.

Moreover, additional cost may be needed to scale
content up. This patch makes wayland window have
app's actual resolution. In webOS, system compositor
scales window to match platform screen size when
window is displayed on screen.

Playing video with webOS media stack it is needed to
provide the target position for the video in UI
resolution (and indicate if it is fullscreen or not).
For that, the algorithm uses additional contents
scale, to adapt the media resolution to the UI
resolution. This is needed for properly showing
fullscreen videos on videos that override the
application resolution at fullscreen (regular way to
see apps in webOS). The window will get the
overridden resolution, but it is upscaled to show
fullscreen. Videos need to adapt to the expected
resolution.

Sometimes on relaunching, the wayland window (and app
size) was resized to display screen size, after
fullscreen. There were two problems. One was that on
setting the window fullscreen, the bounds were set to
the size of the display. This has been fixed
implementing the ToggleFullscreenWithSize method.

The other problem was that, once application got
fullscreen status, window manager was sending a
configure event with the display size. Solution is -
ignoring this. For that this time we use the
WidgetDelegate::CanResize(), that is set to false in
app runtime already. Additionally this patch adds the
handling of the resize behavior that is populated
using CanResize value.

:Testing Performed:
Launch the Viewster app or other 1280x720 app on a
1920x1080 display and check the user interface is
properly shown.

:QA Notes:
Test-cases: QWT-531, QWT-532, QWT-604, QWT-706

:Issues Addressed:
[WRN-2612]   Port [op][n_upsable][dynamic_sized_window]
             Implement dynamic sized window
[PLAT-98783] [chromium79][Signage] Full screen display
             failed when running 1280*720 app.

Change-Id: I21c607f04fb5367b720fae645ec544d5c7663f7f
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/315953
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/317632
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>