[op][upsable][a11y] Allow to make groups of checkboxes
:Release Notes:
Consider groups of checkboxes as valid ARIA sets.

:Detailed Notes:
ARIA does not expect any container with items inside to be a set. It
specifies certain rules for what can be a set, and what kind of items
can be their members. That does not include a "group" of "checkboxes".

This is a bit inconsistent, as on the other hand a menu checkbox is
considered as a valid item inside a menu bar or menu. This change
allows checkbox to be a valid item (a member of a set), and allows
it to be a member of a group set.

:Testing Performed:
1. Open enyo2sampler
2. Go to Moonstone/Checkbox item case
3. Focus an item on the checkbox item group panel
TTS should say the name of the member, if it is checked or not, and
the index and set size (i.e. "Group Option 1, unchecked, 1 of 5").

:QA Notes:
Test-cases: GT-22314

:Issues Addressed:
[PLAT-102530] [A11Y] Check box item group audio guidance does not add count suffix
[WRN-4341] Port : webOS a11y

Change-Id: I25a06b9f1d4297280af807a3ac175544f2c807e8
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/316131
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/317651
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>