[op][ds][build] Add a new build argument for webOS submission
:Release Notes:
Add a new build argument for webOS submission number

:Detailed Notes:
Add a new build argument, named `WEBOS_SUBMISSION_NUMBER` which is set
to webOS submission number. And set a default value of it.

This number is used for versioning of v8. The renewal version of v8 has
effects on metadata of v8 snapshot, codecache and wawm which decides
whether to serialized or not.

note:
It seems more accurate to use v8 submission number rather this
number. Need to review again after a recipe of v8 will be stable.

:Testing Performed:
After build complete, open the webruntime-[version]/Release/args.gn
and check if the following word exists, `webos_submission_number`.

:QA Notes:
The value of `webos_submission_number` can be set to webos metadata, if
not, the default value is set to 0.

:Issues Addressed:
[WRN-4361] Port : [op][n_upsable][build] Add a new build argument for
webOS submission
[QEVENTTWON-4104] WAM is crashed when launching BBC IPlayer

Change-Id: I951868bc7c7bc8b3cca69ac1b47ac8e6e1cfb1c6
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/315737
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/317626
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>