[op][n_upsable][ssl] Allow to use an additional certificates database.
:Release Notes:
Allow to use an additional NSS certificates database for self signed
certificates.

:Detailed Notes:
In webOS a system-wide NSS read only database is used for certificates.
But in certain products, it is interesting to be able to maintain an
additional database for storing self-signed certificates added by the user.

This patch provides a new command line parameter for adding an NSS
database at an specified path: --neva-certificates-path. When passed, Chromium
will also load it (as an NSS user certificate database).

:Testing Performed:
1. Create certificate database path: mkdir /tmp/TestCertDB and set
permissions for WAM.
2. Create a certificates database: certutil -N -d /tmp/TestCertDB
3. Add a badssl certificate to the database: certutil -A -d
/tmp/TestCertDB -n badssl_crt_cer -t "pC,," -i badssl.crt.cer
4. Change the permission of files under /tmp/TestCertDB
chmod 777 /tmp/TestCertDB/*
5. Modify WAM to launch with --neva-certificates-path=/tmp/TestCertDB
and restart it.
6. Sideload https://self-signed.badssl.com -> It should load instead of
showing a certificate error.

:QA Notes:
Test-cases: QWT-4176

:Issues Addressed:
[WRN-4607] Port [op][n_upsable][ssl] Allow to use an additional certificates database.

Change-Id: Ib1455ff9e57dbed130cac4ad7147d0202b88eaaf
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/315688
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/317631
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>