[op][n_upsable][crypto] Provide PAL abstraction for OSCrypt
:Release Notes:
Provide PAL abstraction for OS strings encryption.

:Detailed Notes:
OSCrypt API is used by Chromium to delegate on system encryption, or
provide a strong replacement. In Linux, it uses software encryption
delegating on OS for storing the master key.

In some platforms NEVA should provide support for OS-specific hardware
backed encryption capabilities. This patch implements a PAL abstraction
for that, and exposes it through mojo. This abstraction is then
remotely invoked in NetworkContext

:Testing Performed:
1. Enable logging by adding below command-line flag
  --vmodule=sqlite_persistent_cookie_store=1,*os_crypt*=1
2. Launch any webapp and sideloading http://www.whatarecookies.com/cookietest.asp.
3. Check logging that "Encrypt Cookie" is printing.

:QA Notes:
Test-cases: QWT-4173

:Issues Addressed:
[WRN-4334] Port : [op][n_upsable][crypto] Provide PAL abstraction for OSCrypt
[PLAT-128908] PAL interface for OS encryption

Change-Id: I5adbca48554856010559affe9b3269406269b327
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/315538
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/317630
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>