[op][n_upsable][wayland] Add dummy handler for wl_output::scale
:Release Notes:
Add a handler for wl_output::scale.

:Detailed Notes:
This adds a dummy handler for wl_output::scale event which is required
in the version 2. Otherwise it will crash when receiving that event.
The issue was identified long time ago but has been avoided by LSM to
not send a scale event. That workaround is going to be dropped from LSM
and thus at least a dummy handler should be added to avoid a crash.
I would leave the implementation of the handler to Blink team. For now
it would be okay to leave it unimplemented since it is assumed that the
scale factor is always 1 in webOS.

:Testing Performed:
Built LSM that sends scale event for each wl_output update.
Checked if WAM doesn't crash anymore upon receiving scale event.

:QA Notes:

:Issues Addressed:
[PLAT-131263] Support wl_output@scale in webos-ozone
[WRN-6489] Prepare OSE build with chromium91

Change-Id: I5642a70820ca4b56e22e9a7ed018576365e0a003
Originally-Reviewed-on: http://gpro.lge.com/c/webos-pro/chromium87/+/306156
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/317344
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>