[op][n_upsable][appruntime] New implementation for shifting of overlapped by VKB input field
:Release Notes:
Re-implemented viewport adjustment if an active
input element is overlapped by the activated
virtual keyboard.

:Detailed Notes:
Implemented a new way for shifting of overlapped
by VKB input fields. Content will shift in delta
height of overlapping and will not shift if input
field is going to be shifted out of screen.

:Testing Performed:
1. Launch a WAM-managed app (e.g., UserGuide)
2. Side-load URL for the launched app by
   https://way2tutorial.com/html/example/
   textarea_scrollbar_hide_show_example.php
3. Click any of the textarea input elements available
   on the side-loaded webpage
Expected Result:
Viewport is properly adjusted while any of the textarea input
elements available on the side-loaded page is clicked
so the virtual keyboard shown is about to overlap it

:QA Notes:
Test-cases: QWT-4170

:Issues Addressed:
[WRN-2601]    Port [op][n_upsable][appruntime] New implementation
              for shifting of overlapped by VKB input field
[PLAT-120420] [Auto] Cursor Position is not updated after touching
              second character from AVN VKB

Change-Id: I28a9ab735a9d064966023689401cf5fea5c5f3af
Follows-after: http://gpro.lge.com/315269
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/315316
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/317609
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>