[op][ds][cbe] Do not flag cbe or app_shell as testonly
:Release Notes:
Do not flag cbe or app_shell as testonly, and remove test only code.

:Detailed Notes:
When using enactbrowser, app_shell is our production browser,
so it should not have any test library or component.
So we need to modify the original app_shell testonly flags to be unset.
Code is modified so that a lighter version of content_shell_lib
(named content_shell_nontest_lib) can be used from app_shell.
We also remove the testonly flag of //extensions:shell_and_test_pak.

:Testing Performed:
Build. Use strings on libcbe.so to check "Death test: " is not present.
The string is taken from testing::internal::DeathTestImpl::Passed()
in //third_party/googletest/src/googletest/src/gtest-death-test.cc.

:QA Notes:
Test-cases: QWT-4169, QWT-4168

This is not reducing size back to 70M as we had in chromium68.
Mojoification has made Chromium bigger in general.
Though, with this change we drop around 150K
that is the implementation of the content shell test components.

In chromium 79, libcbe.so size is around 80MB after applying this.
In chromium 84 and 87, it is around 100MB.

:Issues Addressed:
[PLAT-81235] [chromium72] libcbe.so disk size is bigger than 70MB
[WRN-4340] Port : [op][ds][cbe] Do not flag cbe or app_shell as testonly

Change-Id: I20ef70d741d4139d46bf9a36a6e828fbc872c120
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/315329
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/317616
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>