[op][upsable][scroll] Clamp scroll offsets passed to CC from impl side
:Release Notes:
Clamp scroll offsets passed to CC from impl side.

:Detailed Notes:
There is a scenario where, on scrolling fast and close to the limits of
the scrollable area bounds, it could end up generating scroll offsets
out of the scrollable bounds (negative values, or bigger than maximum
scroll offset).

Once that happens, native scroll starts to fail, as it cannot handle
scrolls offset out of the bounds. So, in case the scrollable contents
would be assigned to the layer with wrong scroll_offset (or any child of
it), it ends up failing to scroll, and CC may be confused and draw the
scrolled layer at an unexpected position.

In the detected issue, the scenario was visible when introducing
will-change: transform (or an actual 2d transformation) as it would
make the items be in the scrollable layer.

This scenario was independent of using webOS native scroll (though it
is easier to reproduce with it as scrolling is faster).

It is not clear where this situation is introduced. But we can avoid
the problem just clamping the value when it is set from impl side.

:Testing Performed:
1. Plug an USB disk with a number of photos that is not very big in top
  folder (20-30 is OK).
2. Open Photo&Video
3. Go to USB disk
4. In web inspector, select the node that has the voice intent "Scroll"
   and is parent of the list of items in screen. Add to that node the
   CSS line "will-change: transform;"
5. Scroll up and down to the top repeatedly with remote wheel.

:QA Notes:
Test-cases: QWT-498

:Issues Addressed:
[PLAT-85038] will-change:transform cause incorrect layout when scrolling item in music app list
[WRN-4608] Port [op][upsable][scroll] Clamp scroll offsets passed to CC from impl side

Change-Id: Iccf69d4e62a0d074443709c645cdaf77ffc8c83b
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/317282
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/317670
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>