[op][ds][graphics] Fixed not painting inserted element into the empty layer
:Release Notes:
Fixed painting of an element inserted into
the empty layer.

:Detailed Notes:
When a html element is added to the empty
layer, it doesn't update compositing inputs
and fails to update |draws_content_| of
|GraphicsLayer|, therefore it is skipped for
painting since |draw_contents| is remained
as false.

This is a regression of https://crrev.com/582337,
so this patch reverts part of the patch.

Also this bug is not seen with
compositing-after-paint so once upstream uses
it by default we can drop this patch.

Previous patch didn't work in Chromium87, so
this patch would be on behalf of it.

:Testing Performed:
Load below html content and it should show
the red box on the page.

<html>
<title>Adding a child to an empty layer</title>
<style>
.empty-layer {
  transform: rotateY(30deg) rotateX(-30deg);
  width: 100px;
  height: 100px;
}
.inserted {
  width: 50px;
  height: 50px;
  background-color: red;
}
</style>
<!--
  Force to create composited layer with empty
  then inserts a child the layer.
-->
<div id="empty-layer" class="empty-layer"></div>
<script>
    setTimeout(() => {
      var div = document.getElementById("empty-layer");
      div.innerHTML =
      '<div id="inserted" class="inserted"></div>';
    }, 1000);
</script>

:QA Notes:
Test-cases: QWT-4209

:Issues Addressed:
[WRN-2630]    Port [op][ds][graphics] Fixed not painting
              inserted element into the empty layer
[PLAT-125165] Scheduler no signal icon not displayed

Change-Id: Ia7e41977811f56414d3433b8e70dd7a26ef236de
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/316305
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/317644
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>