[op][upsable][suspend] Do not crash destroying a page if ScopedPageSuspend is set.
:Release Notes:
Avoid certain crashes detected on closing a web application.

:Detailed Notes:
In the destructor of AppRuntimeRenderFrameObserver, we may be
destroying the owned ScopedPagePauser in case it is set. If the frame
observer is destroyed in the handler of Page destruction
(Page::WillBeDestroyed) then we can get a crash in LocalFrame, as
SetContextPaused will use frame_scheduler_, that will be already reset
at that point.

To prevent that situation, we only use the frame scheduler if the frame
is not detached.

:Testing Performed:
Stability 15 hours test.

:QA Notes:
Test-cases: QWT-3872

:Issues Addressed:
[WRN-4480] Port : [op][upsable][suspend] Do not crash destroying a page
if ScopedPageSuspend is set.
[WRN-3564] [RDX_CRASH] /usr/bin/WebAppMgr

Change-Id: Ia768831598a5d644d07763040dff8a5c6d7c137c
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/315884
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/317620
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>