[op][n_upsable][web_security] Do not throw exceptions on illegal local files
:Release Notes:
Allows to start settings application despite it tries access non-existent file

:Detailed Notes:
After RP migration to new chromium v.91 the settings application can't
be started because it tries to access non-existent file. It happens
because web engine throws exception on such illegal access and
application doesn't handle it. In previous versions of chromium
exception was not thrown for local files but aftewards this logic was
removed from upstream chromium. Webruntime kept this behavior in
chromium v.87 by reverting upstream related patch, referred change:
http://gpro.lge.com/c/webos-pro/chromium87/+/301697.
But Neva changed logic of the access restrictions for requests in
chromium v.91 in response to new chromium upstream changes and because
of that RP also needs to adapt the throwing exceptions procedure to
support the work of existing applications.

:Testing Performed:
1. Verify that settings application is started well
Or:
2. From JS console of a trusted app:
   req = XMLHttpRequest();
   req.open("GET", "file:///etc/");
   req.send();
   It should not throw an exception.

:QA Notes:
Test case: QWT-575

:Issues Addressed:
[WRN-4750] [chromium_91] Settings application doesn't start

Change-Id: Id57d4e89c4389a914f464a3b5deda26e391fd839
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/315107
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/317641
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>