[op][ds][appruntime][webos] Restore can_renderer_initiate_transfer logic
:Release Notes:
This CL allows the application to run in the same process after navigation,
this keeps all dedicated parameters for this application unchanged

:Detailed Notes:
When executing the YouTube web app for test,
a new renderer process is created when the app navigates from
file:///usr/palm/applications/com.webos.app.test.youtube/index.html to
https://www.youtube.com/tv, and is_app_preload_hint_set_ of
content::RenderViewImpl is not propagated
from the first renderer process to the second.

This patch reverts the upstream change of
https://chromium-review.googlesource.com/c/chromium/src/+/1867031
so that the app uses the same site instance and renderer process
for both sites.

:Testing Performed:
1. Preload the YouTube web app for test as follows:
  luna-send -n 1 luna://com.webos.applicationManager/launch '
  {"id": "com.webos.app.test.youtube", "launchHidden": true,
  "preload": "partial", "noSplash": true}'
2. Check no cache file is generated.

:QA Notes:
Test-cases: QWT-708, QWT-571, QWT-504

:Issues Addressed:
[WRN-4602] Port [op][ds][appruntime][webos] Restore can_renderer_initiate_transfer logic
[PLAT-126695] Able to scroll the 'HTML' page which is loaded via web-inspector.

Change-Id: I50d4bae5eb5fc4236d97be848d32a502eb7f2e9c
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/315232
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/317607
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>