[op][ds][app_runtime][error_page] Disable error page site isolation
:Release Notes:
Propagation of preferences to error page

:Detailed Notes:
Neva app runtime is still not ready for site isolation or error page
site isolation. We found that at least the WebkitPreferences are not
propagated to the RenderView created for error page.

Upstream policy has changed from disabling by default on content API
the site isolation for error page, to enable it if the error page
replaces the main frame. And the WebkitPreferences in app_runtime::WebView
were not updated accordingly.

As we do not know if there are other side effects we do not know
for enabling error page site isolation, we disable it for now as
we did for general site isolation.

:Testing Performed:
1. restart webapp-mgr
2. Launch viewster, check the PID (the only WebAppMgr with type=renderer).
3. Connect through web inspector and side load a non existing web page
  so it shows the error page
4. Check the only WebAppMgr process with type=renderer keeps being
  the same PID.

:QA Notes:
Test-cases: QWT-458

:Issues Addressed:
[WRN-4601] Port [op][n_upsable][app_runtime][error_page] Disable error page site isolation
[PLAT-117327] Disable error page site isolation

Change-Id: Ie4d47d350e9bf30f9d95c95f9647fe81dc14f279
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/315344
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/317619
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>