[build] Allow ARM 32 bits debug build.
:Release Notes:
Make it possible to build with full debugging information on ARM 32bit platforms.

:Detailed Notes:
This adds several changes:
- Allows to use optimize_for_size GN argument (we should explore just enabling
it for linux if symbol_level is 2).
- We add long-plt for linking CBE.

It seems to require also debug fission or debug compression. It is more practical
to use fission as it reduces build time heavily (compression is a very slow
procedure).

:Testing Performed:
Build completes for ARM32 with symbol_level=2, optimize_for_size=true
and use_debug_fission=true

:QA Notes:

:Issues Addressed:
[OKT-2056]  webOS -g2 compile error
[WRN-4373] Port : [op][upsable][build] Build: fix ARM debugging with symbol_level=2

Change-Id: I874bfa2fbfe1d83020555980f37d8ea2299331ba
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/316387
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/317660
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>