fixup! [op][n_upsable][error_page] Show error page only if error is in main frame
:Release Notes:
Prevent use-after-free and crashes deciding how to handle an error
loading a URL.

:Detailed Notes:
This patch aligns better with upstream expectations for the API
implementation of WebContentsObserver.

Upstream calls DidFailLoad indirectly from RenderFrameHostImpl, so it
is expected the render_frame_host pointer is always valid. So first,
our own handling of DidFailLoad will assert in case it is null.

Then, the only situation that was passing a null pointer was when it
was called through DidFinishNavigation. This was a bad idea, reusing
this way. The new implementation adds a helper method DispatchNetError,
that is called both from DidFailLoad and DidFinishNavigation. In both
sides a proper check for main frame is done (that is different in each
case).

This should fix the hard to reproduce assert that happens when
navigation is canceled before navigation is committed.

:Testing Performed:
Test 1
1. Launch a slow webpage
2. window.close() before it finishes loading.
No crash.

Test 2
1. Set invalid DNS
2. Open Viewster
Error page should be visible.

Test 3
(with proper networking set up)
1. Side load youtube from enact sandstone demo.
2. Navigate to a video and play.
No error page is shown.

:QA Notes:
Test-cases: QWT-571

:Issues Addressed:
[WRN-4631] Port [op][n_upsable][error_page] Show error page only if error is in main frame

Change-Id: I27ff3625ead20d4c2446f416a8d91208ce6882bd
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/317071
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/317663
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>