[op][n_upsable][wayland] Do not set EGL_PLATFORM environment variable after threads are created.
:Release Notes:
Prevent crash caused by setting wayland environment variable, that is
a non thread safe operation, while other thread reads the environment
variables.

:Detailed Notes:
The ozone platform initialization sets the environment variable
EGL_PLATFORM=wayland on GPU process/thread side initialization. This
is OK for GPU process. But setenv is not thread safe, and if another
thread is calling getenv while setenv is done, it could lead to a
crash.

To avoid this problem, we move the EGL_PLATFORM setenv call to
PreEarlyInitialization of the ozone platform backend, so it happens
before any thread is spawn.

:Testing Performed:
Restart WAM and launch one web application.

:QA Notes:
Test-cases: stability.

:Issues Addressed:
[WRN-4589] [RDX_CRASH][webos] /usr/bin/app-shell/app_shell in getenv (libc.so.6 + 0x2a368)

Change-Id: Idb6323ca130d21a3ecbebbaef191e48aaf73ca44
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/317284
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/317671
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>