fixup! [op][ds][log][pmlog] Support PmLog
:Release Notes:
Prevent pmlog maximum message overflow by striping the filenames in log
generated from Chromium.

:Detailed Notes:
Partially recover old behavior for the file name so, for long file names
we don't risk to pass the maximum limit of the PmLog messages. We already
limit the maximum size of the log message itself, but that assumes the
filename will not pass the 896 bytes limit. So we need to keep the 20
characters limit for the file name.

This recovers the previous behavior where we show as much as possible of
the path, not only the filename, providing we keep the 20 chars limit.

:Testing Performed:
Check in PmLog that wam.log channel messages show the filename
in the expected format.

:QA Notes:
Test-cases: QWT-536

:Issues Addressed:
[WRN-6649] /var/log/messages has been replaced; following end of new file error displayed

Change-Id: Ib7f68d6b249e5271d40587072df598edcb622be3
Originally-Reviewed-on: http://gpro.lge.com/c/webos-pro/chromium87/+/317732
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/319295
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Test BTS <test.bts@lge.com>
(cherry picked from commit 9b3ec9c39d376541c9781e42a6c7154748300073)
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/319560
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>