[oe][ds][media][ls2] Match URIs array with URITypes
:Release Notes:
Match URIs array with URITypes

:Detailed Notes:
luna_service_uris array is indexed based on URIType, but there was
a mismatch between the sizes of each.
As some services are not available in OSE, use empty strings so that
there's no buffer overrun when accessing it based on any URIType.

:Testing Performed:
Verified that Media events from bluetooth headset are received from
MCS and the same are processed by the player. This ensures that the
service URI used to register with MCS is proper.

:QA Notes:
Test-cases: QWO-1584

:Issues Addressed:
[WRN-9290] Port [oe][ds][media][ls2] Match URIs array with URITypes

Change-Id: I4786aa8deee45019b22d720386bba3042aa3b2bf
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium87/+/308080
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/318782
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: <andrey.dulevets@lgepartner.com>