[oe][ds][ozone][input] Set the maximum input length to VKB
:Release Notes:
Implemented a way for setting the maximum input length to VKB

:Detailed Notes:
For VKB UI feature:
The Enter button will be disabled if the input field
entered the max allowed characters.
The app will set this max length of the input field and
the imemanger expected to receive from
inputMethodHost()->maxTextLength().
The Enter button will be enabled if the character length is
less than Max input length defined by the apps.

For the feature, in the patch, we get the maximum length
from HTML element and set it to VKB
via text model Wayland interface.

:Testing Performed:
RSPi4 webosose build.
Check Wayland log files and check that max input length
value is set to VKB via text model Wayland interface.

:QA Notes:
Test-cases: QWO-1591

:Issues Addressed:
[WRN-9287] Port [oe][ds]Port [ozone][input] Set the
	   maximum input length to VKB
[PLAT-89649]  Disable the Enter button after entering max
              characters

Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium87/+/308694
Change-Id: Ib9d05540161d6c9ab55150f0a880ba32bcaf5be1
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/319282
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>