[op][n_upsable][ozone] Make touch event ID unique with multiple seats
:Release Notes:
Do not mix touch events from different screens.

:Detailed Notes:
Using dual screen, if a touch gesture is started in one of the screens,
and then another gesture is started in the other screen, if they come
from different seats, GestureRecognizer will fail to properly separate
the events, as the touch lock uses the ID obtained from the wayland
touch slot. But this slot ID is unique to each seat, so same slot ID
can come from each seat.

To avoid that problem, we now generate the ID from a unique number
obtained from the slot and the source device ID.

:Testing Performed:
1. Launch an scrollable page in screen 0.
2. Launch an scrollable page in screen 1.
3. Start scrolling up and down on screen 0.
4. While still scrolling up and down on screen 0, start scrolling up
  and down on screen 1.
The scrolls of both screens should be independent. No flickering
should happen.

:QA Notes:
Test-cases: QWO-490

:Issues Addressed:
[WRN-9302] Port [op][n_upsable]Port [ozone] Make touch event ID unique with multiple seats

Change-Id: Iffac461e5b294a5987cbcf6afc25a4b9b9af07cd
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium87/+/314072
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/319480
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>