[op][n_upsable][lttng] Avoid symbol clash with perfetto and LTTng.
:Release Notes:
Fix building webOS PmTrace support with recent Chromium implementation.

:Detailed Notes:
In Chromium m91, Perfetto libraries are used even if it the client library
is not used, because of the tracing library implementation. But Perfetto
library defines the some properties with the name tracepoint, that clashes
with the LTTng tracepoint define.

To avoid that problem, we rename those to be named trace_point. This is
kind of a work-around, but because LTTng expects direct access to that
define it is just easier to not use that symbol from Perfetto.

:Testing Performed:
1. Build enabling LTTng in WAM and Chromium.
2. Check that WebAppMgr exposes its tracepoints using lttng list -u

:QA Notes:
Test-cases: QWT-371

:Issues Addressed:
[WRN-12099] Strings are not displayed in Trace-file.

Change-Id: I91832d748749dd1e2bd56c13458be5e6b15d2bab
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/321834
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/324773
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>