[op][n_upsable][cursor] Fix crash processing CSS cursor rules fot webOS
:Release Notes:
Fix certain knobs crashing on being dragged.

:Detailed Notes:
The hack added to have specific processing of cursor rules for webOS,
available from NEVA, is fetching Element style rulesets in a non-safe
way, as at the point they are accessed, they could be invalid.

To avoid that, we force a layout rules update. This is not ideal, and
the right way to fix this is through computed styles.

:Testing Performed:
Try several slides knobs from latest Enact Sandstone, to see if they
do not crash.

:QA Notes:
Test-cases: QWT-2528, QWT-2694, QWT-2741, QWT-2744, QWT-2780

:Issues Addressed:
[WRN-12879] Progress knob doesn't slide fully to extreme right & Crash is observed 

Change-Id: I8cd8ede6663c8a1e17a4c1308b04938f889df30f
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/322028
Tested-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/325957
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>