[op][upsable][a11y] Restore original parents of aria-owns owned children when owner is removed
:Release Notes:
Fix losing part of the accessible readout when setting
aria-owner, then removing it.

:Detailed Notes:
When aria-owns relationship is set, the owned children accessible
parent is set to the owner. But, if the owner is removed, the
relationship should be reset to the original parent.

This is an implementation that is specific to chromium91. An
equivalent implementation should be possible in chromium87.

In upstream, implementation has changed so a new implementation
is being proposed at:
https://chromium-review.googlesource.com/c/chromium/src/+/3386523

The upstream bug is being tracked at:
https://bugs.chromium.org/p/chromium/issues/detail?id=1287181

:Testing Performed:
1. Create an element that aria-owns a different part of the DOM
tree that has an accessible representation (i.e. with labelled
buttons).
2. Remove that element.
3. Check the accessible representation of the part that was
previously attached through aria-owns is back to its original
place.

:QA Notes:
Test-cases: QT-4440

:Issues Addressed:
[WRN-13457] [ch87][accessibility] Do not speak when relaunching and redirect page

Change-Id: I94a7f411b9e375a16838ab8e22559491c6354b88
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/324963
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/325959
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>