[oe][ds][gpu] Bypass the error when shared image is not cleared
:Release Notes:
Workaround for bypassing the error when shared image is not
cleared as it's supposed to be

:Detailed Notes:
In rare cases we get error when shared image is not cleared as
expected. The root of this problem sides in
RasterDecoderImpl::DoCopySubTextureINTERNALGL where the texture
size does not match the expected copied rectangle, and this leads
to the fact that the 'clear' attribute is not set for backing
storage. This is workaround where we are trying to continue work
and not fall into error when such problem is met.

:Testing Performed:
1. Power off RPi board
2. Power on RPi board
3. Verify that HOME app ribbon and overall screen has no visible
artefacts.
4. Repeat step 3 several times

:QA Notes:
Test-cases: QWO-1

:Issues Addressed:
[WRN-16831] App titles and icons are not displayed on screen

Change-Id: I823e3178d67b45a2584bfb1492252c9f76fbad64
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/326988
Tested-by: Vladislav Mukulov <vladislav.mukulov@lge.com>
Tested-by: Test BTS <test.bts@lge.com>
Reviewed-by: Vladislav Mukulov <vladislav.mukulov@lge.com>
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>