fixup! [op][ds][media] Implement MediaPlatformAPIWebOSGmp
:Release Notes:
Discard feeding old audio buffers to g-media-pipeline

:Detailed Notes:
When you re-launch YouTube application i.e onResume(), sometimes audio
is not heard because of overflow.

This happens because, sometimes during resume of a suspended player,
in MediaPlatformAPIWebOSGMP we receive old audio buffers with time
stamp much ahead of the current resume time. So GMP is not able to
consume these buffers and a discontinuity and overflow situation
occurs.

With this patch, we are flushing the old buffers whenever app goes to
the suspend state. In this way we are able to discard the old buffers
without feeding them to GMP. So the overflow situation does not occur.

:Testing Performed:
1. Launch YouTube and play any video
2. Launch com.webos.app.test.webosose test app.
3. Re-Launch YouTube
4. Repeat 1-3, Youtube video should play with audio

:QA Notes:
Test-cases: QWO-489

:Issues Addressed:
[WRN-19158] [webOS OSE]: Audio is not heard when youtube Re-Launch

Change-Id: I1d717f02c4b497f243b20fa8f3c73b576b892314
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/328243
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Praveen Kumar P <praveen.p@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Praveen Kumar P <praveen.p@lge.com>