[op][n_upsable][app_shell] Enable audio and video capture permissions
:Release Notes:
Enabled Audio and Video capture permissions which were blocked in
the older versions.

:Detailed Notes:
Audio and Video capture permission implementation for app_shell was
causing crash in VerifyMediaAccessPermission function in older versions
of chromium.
In http://text.lge.com/neva/gerrit/#/c/2781/ the fix was done to return
false and not access VerifyMediaAccessPermission function.
Reverted this patch set as crash is not observed in the latest versions
of Chromium.

:Testing Performed:
Connect the camera and mic devices
1. Launch Enact Browser
2. Open website: "https://webrtc.github.io/samples/src/content/
   devices/input-output/" or "https://yandex.ru" or
   "https://mictests.com/"
3. Check that Camera and mic devices connected are working and allow
   to select the device when multiple devices are connected.

:QA Notes:
Test-cases: QWTC-8130

:Issues Addressed:
[WRO-9582] Media devices (Camera/Mic) are not listed in Enact Browser

Change-Id: If31624637e9adce73fb3cd33bfd08b33d8f3c94c
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/334387
Reviewed-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Bharath Patil <bharath.patil@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Bharath Patil <bharath.patil@lge.com>