fixup! [op][n_upsable][media] Fix suspend/resume with WebMediaPlayerImpl
:Release Notes:
Refactor suspend/resume implementation for WebMediaPlayerImpl to
support Enact browser as well

:Detailed Notes:
With the use of WebMediaPlayerImpl for all types of Media playback
we have suspend/resume issue remains for Enact browser when the
tab goes to background.

With this patch set we move the MediaSuspender instance creation to
the common class AppRuntimeWebViewControllerImpl, so that the same
code can be used by both web apps and Enact browser app shell. And
suppressed resuming media playback while in background

Similarly moved the MediaPlaybackOptions creation to the common
class AppRuntimeRenderFrameObserver to be used by both.

Additionally, in MediaSuspender we changed the media playback
naming to video playback naming to be inline with naming of render
frame media playback options.

:Testing Performed:
1. Launch Enact browser and load youtube in one tab and play video
2. Open a second tab and load google.com page
3. Video playback should be stopped in first tab with youtube

1. Launch google Media Session Sample page with backgroundmedia app
2. Launch google.com search webpage with test.redirection app
3. Video playback should be continued with backgroundmedia app
4. Close backgroundmedia app and launch youtube app and play video
5. Repeat step-2. Video playback in youtube app should be stopped

:QA Notes:
Test-cases: QWTC-6658, QWTC-6833, QWTC-435 , QWTC-436

:Issues Addressed:
[WRO-11385] HTML5 Video in the first tab continuous to play and
            the sound is heard even after opening second tab

Change-Id: I4b12c9f1ace8142f4ae14edb3efa9b9122e42fa3
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/335128
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Karthick Sundaram <karthick.sundaram@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Pankaj Maharana <pankaj.maharana@lge.com>